[onert/lib] Move CopyInputInitializer class definition
This commit get CopyInputInitializer out of RandomTestRunner.cpp like RandomInputInitializer
